fixed a potential floating point error

+++ mega-cart/cart-service/carts.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS items (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255),
     description VARCHAR(255),
-    price FLOAT(2),
+    price NUMERIC(10, 2),
     quantity INT,
     cart_id INTEGER REFERENCES carts(id) ON DELETE CASCADE
 );
